+++ lib/services/file_access_service_web.dart
@@ -66,6 +66,7 @@ class FileAccessService {
   }
 
   Future<void> evictCacheForObject({
+    required RemoteStorageConfig config,
     required String bucket,
     required ObjectInfo object,
   }) async {}